pyproject.toml
[tool.poetry]
name = "crudProj"
version = "0.1.0"
description = "CRUD api in python"
authors = ["Shubh Srivastava"]
readme = "README.md"
keywords = ["CRUD", "Python", "Database"]

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "^8.1.0"
pytest = "^7.4.2"
pylint = "^2.17.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
